+++ kpoint-react/src/components/all-projects-page/project-card.tsx
@@ -1,6 +1,5 @@
 import ArrowForwardTwoToneIcon from '@mui/icons-material/ArrowForwardTwoTone';
 import BookmarkTwoToneIcon from '@mui/icons-material/BookmarkTwoTone';
-import ControlPointTwoToneIcon from '@mui/icons-material/ControlPointTwoTone';
 import MonetizationOnTwoToneIcon from '@mui/icons-material/MonetizationOnTwoTone';
 import PeopleAltTwoToneIcon from '@mui/icons-material/PeopleAltTwoTone';
 import ShareTwoToneIcon from '@mui/icons-material/ShareTwoTone';
@@ -15,14 +14,12 @@ import Grid from '@mui/material/Grid';
 import IconButton from '@mui/material/IconButton';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
-import { FC, useEffect, useState } from 'react';
+import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
-import { toast } from 'react-toastify';
-import { projectAction } from 'store/actions';
 
 import { StorageKey } from '../../common/enums/app/storage-key.enum';
-import { useAppDispatch } from '../../hooks/use-app-dispatch/use-app-dispatch.hook';
 import { storage } from '../../services/services';
+import { SubscribeButton } from './SubscribeButton';
 
 interface ProjectsProps {
   projectId: string;
@@ -36,27 +33,7 @@ interface ProjectsProps {
 const ProjectCard: FC<ProjectsProps> = ({ projectId, url, title, summary,
   logoImgUrl, tags }) => {
   const { t } = useTranslation();
-  const [isFollowing, setIsFollowing] = useState(false);
-  const dispatch = useAppDispatch();
   const userID = storage.getItem(StorageKey.USER);
-
-  useEffect(() => {
-    const isUserSubscribed = localStorage.getItem(`project_${projectId}_${userID}_subscription`);
-
-    if (isUserSubscribed) {
-      setIsFollowing(true);
-    }
-
-  }, [projectId, userID]);
-
-  const handleButtonSubClick = (): void => {
-    if (!isFollowing) {
-      setIsFollowing(true);
-      localStorage.setItem(`project_${projectId}_${userID}_subscription`, 'true');
-      dispatch(projectAction.subscribeToProject({ projectId: projectId }));
-      toast.success('Ви успішно підписані на проект');
-    }
-  };
 
   return (
     <Card sx={{ maxWidth: 500 }}>
@@ -76,10 +53,7 @@ const ProjectCard: FC<ProjectsProps> = ({ projectId, url, title, summary,
             aria-label="outlined button group" sx={{ margin: 1 }}>
             <Button size="small" startIcon={ <PeopleAltTwoToneIcon/> }
               sx={{ justifyContent: 'right' }}>{t('buttons.help')}</Button>
-            <Button size="small" startIcon={ <ControlPointTwoToneIcon/> }
-              sx={{ justifyContent: 'right' }}onClick={handleButtonSubClick} >
-              {isFollowing ? t('buttons.unfollow') : t('buttons.follow')}
-            </Button>
+            <SubscribeButton projectId={projectId} userID={userID ? userID : ''} />
             <Button size="small" startIcon={ <MonetizationOnTwoToneIcon/> }
               sx={{ justifyContent: 'right' }}>
               {t('buttons.donate_projects_page')}</Button>
